replace formatDuration with formatDurationLong and move to utils.formatting.time
