--- frontend/src/metabase/query_builder/components/view/ExecutionTime/utils.ts
@@ -1,9 +1,0 @@
-import { t } from "ttag";
-
-export const formatDuration = (time: number): string => {
-  if (time < 1000) {
-    return t`${time}ms`;
-  }
-
-  return t`${(time / 1000).toFixed(1)}s`;
-};
