+++ frontend/src/metabase/transforms/utils.ts
@@ -1,4 +1,3 @@
-import dayjs from "dayjs";
 import { t } from "ttag";
 import _ from "underscore";
 
@@ -274,47 +273,6 @@ export const getRootCollectionItem = ({
   }
   return null;
 };
-
-/**
- * Human-readable duration formatter tuned for transform run lengths.
- *
- * Ladder:
- *   < 1s     → "Nms"    (e.g. "750ms")
- *   < 1min   → "N.Ms"   (e.g. "8.4s")
- *   < 1hr    → "Xm Ys"  (e.g. "8m 42s")
- *   ≥ 1hr    → "Xh Ym"  (e.g. "2h 17m")
- *
- * Negative inputs (defensive — clock skew between FE and DB) clamp to 0ms.
- *
- * Uses dayjs.duration internally for component decomposition (the duration
- * plugin is already extended globally at frontend/src/metabase/utils/dayjs.ts),
- * but NOT its .format("HH:mm:ss") because that produces clock-style output
- * rather than the readable ladder we want.
- */
-export function formatDurationLong(durationMs: number): string {
-  if (durationMs <= 0) {
-    return t`0ms`;
-  }
-  if (durationMs < 1_000) {
-    const ms = Math.round(durationMs);
-    return t`${ms}ms`;
-  }
-  if (durationMs < 60_000) {
-    const seconds = (durationMs / 1_000).toFixed(1);
-    return t`${seconds}s`;
-  }
-  const d = dayjs.duration(durationMs);
-  if (durationMs < 3_600_000) {
-    const minutes = d.minutes();
-    const seconds = d.seconds();
-    return t`${minutes}m ${seconds}s`;
-  }
-  // Hour-plus: use TOTAL hours (asHours can exceed 24 for very long
-  // backfills) plus the leftover minutes within the hour.
-  const hours = Math.floor(d.asHours());
-  const minutes = d.minutes();
-  return t`${hours}h ${minutes}m`;
-}
 
 /**
  * Returns the duration in ms of a transform run, or null when it cannot be
