+++ frontend/src/metabase/utils/formatting/time.ts
@@ -1,5 +1,5 @@
-import type { Dayjs } from "dayjs";
-import { msgid, ngettext } from "ttag";
+import dayjs, { type Dayjs } from "dayjs";
+import { msgid, ngettext, t } from "ttag";
 
 import type { TimeOnlyOptions } from "metabase/utils/formatting/types";
 import { parseTime, parseTimestamp } from "metabase/utils/time-dayjs";
@@ -27,6 +27,47 @@ export function duration(milliseconds: number) {
   const seconds = Math.round(milliseconds / SECOND);
 
   return ngettext(msgid`${seconds} second`, `${seconds} seconds`, seconds);
+}
+
+/**
+ * Human-readable duration formatter.
+ *
+ * Ladder:
+ *   < 1s     → "Nms"    (e.g. "750ms")
+ *   < 1min   → "N.Ms"   (e.g. "8.4s")
+ *   < 1hr    → "Xm Ys"  (e.g. "8m 42s")
+ *   ≥ 1hr    → "Xh Ym"  (e.g. "2h 17m")
+ *
+ * Negative inputs (defensive — e.g. clock skew between FE and DB) clamp to 0ms.
+ *
+ * Uses dayjs.duration internally for component decomposition (the duration
+ * plugin is already extended globally at frontend/src/metabase/utils/dayjs.ts),
+ * but NOT its .format("HH:mm:ss") because that produces clock-style output
+ * rather than the readable ladder we want.
+ */
+export function formatDurationLong(durationMs: number): string {
+  if (durationMs <= 0) {
+    return t`0ms`;
+  }
+  if (durationMs < 1_000) {
+    const ms = Math.round(durationMs);
+    return t`${ms}ms`;
+  }
+  if (durationMs < 60_000) {
+    const seconds = (durationMs / 1_000).toFixed(1);
+    return t`${seconds}s`;
+  }
+  const d = dayjs.duration(durationMs);
+  if (durationMs < 3_600_000) {
+    const minutes = d.minutes();
+    const seconds = d.seconds();
+    return t`${minutes}m ${seconds}s`;
+  }
+  // Hour-plus: use TOTAL hours (asHours can exceed 24 for very long
+  // backfills) plus the leftover minutes within the hour.
+  const hours = Math.floor(d.asHours());
+  const minutes = d.minutes();
+  return t`${hours}h ${minutes}m`;
 }
 
 interface TimeWithUnitType {
